Chapter08/Code/requirements.txt
absl-py==0.3.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2018.10.15
cloudpickle==0.6.1
configparser==3.5.0
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gast==0.2.0
graphviz==0.8
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
imageio==2.4.1
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
linecache2==1.0.0
Markdown==2.6.11
matplotlib==2.2.2
mkl-fft==1.0.4
mkl-random==1.0.1
mock==2.0.0
more-itertools==4.3.0
multipledispatch==0.6.0
mxnet==0.11.0rc1
networkx==2.2
numpy==1.14.5
olefile==0.45.1
pandas==0.23.2
pathlib2==2.3.2
patsy==0.5.0
pbr==4.2.0
Pillow==5.2.0
pluggy==0.7.1
protobuf==3.6.1
py==1.6.0
pyparsing==2.2.0
pyro-ppl==0.2.0+9c3d68e
pytest==3.8.1
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.13
scandir==1.9.0
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
singledispatch==3.4.0.3
six==1.11.0
statsmodels==0.9.0
subprocess32==3.5.2
tensorboard==1.11.0
tensorflow==1.11.0
tensorflow-probability==0.4.0
termcolor==1.1.0
toolz==0.9.0
tornado==5.1
tqdm==4.23.4
traceback2==1.4.0
unittest2==1.1.0
Werkzeug==0.14.1
wget==3.2
